only creates songs for test and dev, also fixes content-security-policy

+++ src/index.ts
@@ -3,6 +3,7 @@ import dotenv from 'dotenv';
 import Debug from 'debug';
 import supportsColor from 'supports-color';
 import express from 'express';
+import type { Request, Response } from 'express';
 import mongoose from 'mongoose';
 import helmet from 'helmet';
 import bodyParser from 'body-parser';
@@ -37,8 +38,23 @@ if (process.env.NODE_ENV === 'test') mongoDbUri = process.env.TEST_DB || /* ista
 mongoose.connect(mongoDbUri, {
   useNewUrlParser: true, useUnifiedTopology: true, useFindAndModify: false, useCreateIndex: true,
 });
-app.use(helmet({
-  contentSecurityPolicy: false,
+app.use(helmet());
+app.use(helmet.contentSecurityPolicy({
+  directives: {
+    'default-src': ["'self'"],
+    'base-uri': ["'self'"],
+    'block-all-mixed-content': [],
+    'font-src': ["'self'", 'https:', 'data:'],
+    'frame-src': ["'self'", 'https://accounts.google.com'],
+    'frame-ancestors': ["'self'"],
+    'img-src': ["'self'", 'data:', 'https:'],
+    'object-src': ["'none'"],
+    'script-src': ["'self'", 'https://maps.googleapis.com', 'https://apis.google.com'],
+    'script-src-attr': ["'none'"],
+    'style-src': ["'self'", 'https:', "'unsafe-inline'"],
+    'upgrade-insecure-requests': [],
+    'connect-src': ["'self'", 'ws:', 'wss:'],
+  },
 }));
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
@@ -47,7 +63,7 @@ routes(app);
 app.get('*', (req, res) => {
   res.sendFile(path.normalize(path.join(__dirname, '../JaMmusic/dist/index.html')));
 });
-app.use((err: any, req, res: any) => {
+app.use((err: any, req:Request<any>, res: Response<any>) => {
   res.status(err.status || 500)
     .json({ message: err.message, error: err });
 });
@@ -61,15 +77,15 @@ app.use((err: any, req, res: any) => {
     debug(result);
   });
 }
-// if (process.env.NODE_ENV !== 'production') {
-(async () => {
-  const { songs } = songData;
-  try {
-    await songController.deleteAllDocs();
-    await songController.createDocs(songs);
-  } catch (e) /* istanbul ignore next */{ debug(e.message); return Promise.resolve(e.message); }
-  return Promise.resolve('songs created');
-})();
-// }
+if (process.env.NODE_ENV !== 'production') {
+  (async () => {
+    const { songs } = songData;
+    try {
+      await songController.deleteAllDocs();
+      await songController.createDocs(songs);
+    } catch (e) /* istanbul ignore next */{ debug(e.message); return Promise.resolve(e.message); }
+    return Promise.resolve('songs created');
+  })();
+}
 debug(`isTTY?: ${process.stderr.isTTY}`);
 export default app;
